checkpoint - zig build test works

+++ libs/update_modules.py
@@ -8,6 +8,7 @@ import os
 import shutil
 import tempfile
 
+
 def _parse_args():
     """ parse arguments out of sys.argv """
     parser = argparse.ArgumentParser(
@@ -46,16 +47,23 @@ def main():
 
         for d in dirs_to_update:
             destination_path = os.path.join(path_root, d)
+
             to_path_tmp = os.path.join(tmpdir, d)
             print(f"Moving {destination_path} to {to_path_tmp}")
             shutil.move(destination_path, to_path_tmp)
 
             from_path = os.path.join(args.path, "libs", d)
+            if not os.path.exists(from_path):
+                print(
+                    f"WARNING: skipping {from_path}, does not exist in source"
+                )
+                continue
+
             print(f"Copying {from_path} to {destination_path}")
             shutil.copytree(from_path, destination_path)
 
     print("Done.")
-        
+
 
 if __name__ == "__main__":
     main()
